\name{canvas}
\alias{canvas}
\title{
Blank plot
}
\description{
Create a blank canvas with coordinates that go from 0 to 1.
}
\usage{
canvas()
}
\value{
nothing useful.
}
\section{Side effects }{
create a plot on the current graphics device, 
and modify the random seed.
}
\references{
\url{http://www.burns-stat.com/documents/books/tao-te-programming/}
}
\examples{
canvas()
}
\keyword{ hplot }