\name{TaoTeProgramming-package}
\alias{TaoTeProgramming-package}
\alias{TaoTeProgramming}
\docType{package}
\title{
Illustrations from the book 'Tao Te Programming'
}
\description{
The functions to create the actual illustrations plus the 
underlying random-based functions that they rely upon.
}
\details{
\tabular{ll}{
Package: \tab TaoTeProgramming\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-08\cr
License: \tab Unlimited\cr
}
}
\author{
Pat Burns

Maintainer: Pat Burns <patrick@burns-stat.com>
}
\references{
\url{http://www.burns-stat.com/documents/books/tao-te-programming/}
}
\keyword{ package }

